+++ data/mypage/src/main/java/org/sopt/official/data/mypage/remote/model/response/LogOutResponse.kt
@@ -22,31 +22,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.data.source.impl
+package org.sopt.official.data.mypage.remote.model.response
 
-import org.sopt.official.common.di.Auth
-import org.sopt.official.data.model.request.LogOutRequest
-import org.sopt.official.network.model.response.AuthResponse
-import org.sopt.official.data.model.response.LogOutResponse
-import org.sopt.official.data.service.AuthService
-import org.sopt.official.data.source.api.auth.RemoteAuthDataSource
-import org.sopt.official.network.model.request.RefreshRequest
-import org.sopt.official.network.service.RefreshService
-import javax.inject.Inject
+import domain.model.LogOut
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
-class DefaultRemoteAuthDataSource @Inject constructor(
-    @Auth private val authService: AuthService,
-    private val refreshService: RefreshService,
-) : RemoteAuthDataSource {
-    override suspend fun refresh(token: RefreshRequest): AuthResponse {
-        return refreshService.refresh(token)
-    }
-
-    override suspend fun withdraw() {
-        authService.withdraw()
-    }
-
-    override suspend fun logout(request: LogOutRequest): LogOutResponse {
-        return authService.logOut(request)
-    }
+@Serializable
+data class LogOutResponse(
+    @SerialName("success")
+    val success: Boolean,
+    @SerialName("message")
+    val message: String
+){
+    fun toEntity() = LogOut(
+        success, message
+    )
 }
